chore(session): remove unused deprecated fields
