+++ discord/helper_test.go
@@ -1,4 +1,4 @@
-package gokord
+package discord
 
 import (
 	"testing"
