+++ session.go
@@ -13,7 +13,7 @@ import (
 	"github.com/nyttikord/gokord/user/status"
 )
 
-// A Session represents a connection to the Discord API.
+// Session represents a connection to the Discord API.
 type Session struct {
 	sync.RWMutex
 
@@ -41,8 +41,7 @@ type Session struct {
 	ShardCount int
 
 	// Should state tracking be enabled.
-	// State tracking is the best way for getting the users
-	// active guilds and the members of the guilds.
+	// State tracking is the best way for getting the users active guilds and the members of the guilds.
 	StateEnabled bool
 
 	// Whether to call event handlers synchronously.
@@ -51,46 +50,40 @@ type Session struct {
 
 	// Exposed but should not be modified by User.
 
-	// Whether the Data Websocket is ready
-	DataReady bool // NOTE: Maye be deprecated soon
+	// Whether the Data Websocket is ready.
+	//
+	// Note: May be deprecated soon.
+	DataReady bool
 
-	// Max number of REST API retries
+	// Max number of REST API retries.
 	MaxRestRetries int
 
-	// Status stores the current status of the websocket connection
-	// this is being tested, may stay, may go away.
+	// Status stores the current status of the websocket connection this is being tested, may stay, may go away.
 	status int32
 
-	// Whether the Voice Websocket is ready
-	VoiceReady bool // NOTE: Deprecated.
-
-	// Whether the UDP Connection is ready
-	UDPReady bool // NOTE: Deprecated
-
-	// Stores a mapping of guild id's to VoiceConnections
+	// Stores a mapping of guild id's to VoiceConnection.
 	VoiceConnections map[string]*VoiceConnection
 
-	// Managed state object, updated internally with events when
-	// StateEnabled is true.
+	// Managed state object, updated internally with events when StateEnabled is true.
 	State *State
 
-	// The http client used for REST requests
+	// The http.Client used for REST requests.
 	Client *http.Client
 
-	// The dialer used for WebSocket connection
+	// The websocket.Dialer used for WebSocket connection.
 	Dialer *websocket.Dialer
 
-	// The user agent used for REST APIs
+	// The UserAgent used for REST APIs.
 	UserAgent string
 
-	// Stores the last HeartbeatAck that was received (in UTC)
+	// Stores the LastHeartbeatAck that was received (in UTC).
 	LastHeartbeatAck time.Time
 
-	// Stores the last Heartbeat sent (in UTC)
+	// Stores the LastHeartbeatSent (in UTC).
 	LastHeartbeatSent time.Time
 
-	// used to deal with rate limits
-	Ratelimiter *RateLimiter
+	// Used to deal with rate limits.
+	RateLimiter *RateLimiter
 
 	// Event handlers
 	handlersMu   sync.RWMutex
@@ -103,31 +96,30 @@ type Session struct {
 	// When nil, the session is not listening.
 	listening chan interface{}
 
-	// sequence tracks the current gateway api websocket sequence number
+	// sequence tracks the current gateway api websocket sequence number.
 	sequence *int64
 
-	// stores sessions current Discord Resume Gateway
+	// Stores sessions current Discord Resume Gateway.
 	resumeGatewayURL string
 
-	// stores sessions current Discord Gateway
+	// Stores sessions current Discord Gateway.
 	gateway string
 
-	// stores session ID of current Gateway connection
+	// Stores session ID of current Gateway connection.
 	sessionID string
 
-	// used to make sure gateway websocket writes do not happen concurrently
+	// Used to make sure gateway websocket writes do not happen concurrently.
 	wsMutex sync.Mutex
 }
 
-// A TooManyRequests struct holds information received from Discord when receiving an HTTP 429 response.
+// TooManyRequests holds information received from Discord when receiving an HTTP 429 response.
 type TooManyRequests struct {
 	Bucket     string        `json:"bucket"`
 	Message    string        `json:"message"`
 	RetryAfter time.Duration `json:"retry_after"`
 }
 
-// UnmarshalJSON helps support translation of a milliseconds-based float
-// into a time.Duration on TooManyRequests.
+// UnmarshalJSON helps support translation of a milliseconds-based float into a time.Duration on TooManyRequests.
 func (t *TooManyRequests) UnmarshalJSON(b []byte) error {
 	u := struct {
 		Bucket     string  `json:"bucket"`
@@ -146,14 +138,14 @@ func (t *TooManyRequests) UnmarshalJSON(b []byte) error {
 	return nil
 }
 
-// GatewayBotResponse stores the data for the gateway/bot response
+// GatewayBotResponse stores the data for the gateway/bot response.
 type GatewayBotResponse struct {
 	URL               string             `json:"url"`
 	Shards            int                `json:"shards"`
 	SessionStartLimit SessionInformation `json:"session_start_limit"`
 }
 
-// SessionInformation provides the information for max concurrency sharding
+// SessionInformation provides the information for max concurrency sharding.
 type SessionInformation struct {
 	Total          int `json:"total,omitempty"`
 	Remaining      int `json:"remaining,omitempty"`
@@ -161,7 +153,7 @@ type SessionInformation struct {
 	MaxConcurrency int `json:"max_concurrency,omitempty"`
 }
 
-// GatewayStatusUpdate is sent by the client to indicate a presence or status update
+// GatewayStatusUpdate is sent by the client to indicate a presence or status update.
 // https://discord.com/developers/docs/topics/gateway#update-status-gateway-status-update-structure
 type GatewayStatusUpdate struct {
 	Since  int             `json:"since"`
@@ -182,7 +174,7 @@ type Identify struct {
 	Intents        discord.Intent      `json:"intents"`
 }
 
-// IdentifyProperties contains the "properties" portion of an Identify packet
+// IdentifyProperties contains the "properties" portion of an Identify packet.
 // https://discord.com/developers/docs/topics/gateway#identify-identify-connection-properties
 type IdentifyProperties struct {
 	OS              string `json:"$os"`
