+++ wsapi.go
@@ -17,17 +17,14 @@ import (
 	"github.com/nyttikord/gokord/user/status"
 )
 
-// ErrWSAlreadyOpen is thrown when you attempt to open
-// a websocket that already is open.
-var ErrWSAlreadyOpen = errors.New("web socket already opened")
-
-// ErrWSNotFound is thrown when you attempt to use a websocket
-// that doesn't exist
-var ErrWSNotFound = errors.New("no websocket connection exists")
-
-// ErrWSShardBounds is thrown when you try to use a shard ID that is
-// more than the total shard count
-var ErrWSShardBounds = errors.New("ShardID must be less than ShardCount")
+var (
+	// ErrWSAlreadyOpen is thrown when you attempt to open a websocket that already is open.
+	ErrWSAlreadyOpen = errors.New("web socket already opened")
+	// ErrWSNotFound is thrown when you attempt to use a websocket that doesn't exist
+	ErrWSNotFound = errors.New("no websocket connection exists")
+	// ErrWSShardBounds is thrown when you try to use a shard ID that is more than the total shard count
+	ErrWSShardBounds = errors.New("ShardID must be less than ShardCount")
+)
 
 type resumePacket struct {
 	Op   int `json:"op"`
